+++ src/ol_openedx_canvas_integration/__init__.py
@@ -2,6 +2,6 @@
 This is an integration of canvas with edX.
 """
 
-__version__ = "0.0.1"
+__version__ = "0.5.0"
 
 default_app_config = "ol_openedx_canvas_integration.app.CanvasIntegrationConfig"  # pylint: disable=invalid-name