+++ src/ol_openedx_canvas_integration/CHANGELOG.rst
@@ -12,7 +12,13 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
-*
+[0.5.0] - 2025-06-19
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Added
+_____
+
+* Automatically update grades in Canvas if the assignments are synced between Open edX and Canvas.
+* Automatically sync assignments in Open edX to Canvas when the courses are linked.
 
 [0.0.1] - 2021-10-29
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~